requirements.txt
Flask==2.0.1
requests==2.26.0
Pillow==8.3.1
openai==0.6.0
gunicorn==22.0.0
Werkzeug==2.2.2